03-aidd/pyproject.toml
[project]
name = "telegram-film-bot"
version = "0.1.0"
description = "Telegram bot with LLM integration for film expert conversations"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "aiogram>=3.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dependency-groups = {}

